Adds layer as a top-level attribute for models.
Layers used to be rolled in with position data, but doing so introduced a couple of issues. The bigger one was that changing layers was a slow and difficult process and the other that the layer information being included with the rest of the position data was wasteful. The layer never changes, so it is a prime use-case for a uniform in the vertex shader. This change improves both performance and maintainability of the code.